+++ apis/vercel/next.config.js
@@ -3,9 +3,9 @@ const path = require("path");
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   poweredByHeader: false,
-  outputFileTracingRoot: path.join(__dirname, "../../.."),
+  outputFileTracingRoot: path.join(__dirname, "../.."),
   turbopack: {
-    root: path.join(__dirname, "../../.."),
+    root: path.join(__dirname, "../.."),
   },
   transpilePackages: ["@braintrust/proxy", "@vercel/examples-ui"],
   typescript: {
