+++ packages/proxy/package.json
@@ -31,6 +31,12 @@
       "module": "./schema/dist/index.mjs",
       "require": "./schema/dist/index.js"
     },
+    "./schema/translate": {
+      "types": "./schema/translate/dist/translate.d.ts",
+      "import": "./schema/translate/dist/translate.mjs",
+      "module": "./schema/translate/dist/translate.mjs",
+      "require": "./schema/translate/dist/translate.js"
+    },
     "./utils": {
       "types": "./utils/dist/index.d.ts",
       "import": "./utils/dist/index.mjs",
@@ -54,6 +60,7 @@
     "dist/**/*",
     "edge/dist/**/*",
     "schema/dist/**/*",
+    "schema/translate/dist/**/*",
     "types/dist/**/*",
     "providers/dist/**/*"
   ],
