+++ packages/proxy/schema/index.ts
@@ -1,7 +1,6 @@
 import { z } from "zod";
 import type {
   AnyModelParamsType as AnyModelParam,
-  ChatCompletionMessageParamType as Message,
   MessageRoleType as MessageRole,
   ModelParamsType as ModelParams,
 } from "../src/generated_types";
@@ -12,9 +11,6 @@ import {
   ModelSpec,
   getAvailableModels,
 } from "./models";
-import { openaiParamsToAnthropicMesssageParams } from "@lib/providers/anthropic";
-import { OpenAIChatCompletionCreateParams } from "@types";
-import { openaiParamsToGeminiMessageParams } from "@lib/providers/google";
 
 export * from "./secrets";
 export * from "./models";
@@ -75,19 +71,6 @@ export const modelParamToModelParam: {
   reasoning_budget: "reasoning_budget",
   verbosity: "verbosity",
   stop: null,
-};
-
-const paramMappers: Partial<
-  Record<
-    ModelFormat,
-    (
-      params: OpenAIChatCompletionCreateParams,
-      modelSpec?: ModelSpec | null,
-    ) => object
-  >
-> = {
-  anthropic: openaiParamsToAnthropicMesssageParams,
-  google: openaiParamsToGeminiMessageParams,
 };
 
 export const sliderSpecs: {
@@ -1045,58 +1028,6 @@ export const EndpointProviderToBaseURL: {
   databricks: null,
   js: null,
 };
-
-export function buildClassicChatPrompt(messages: Message[]) {
-  return (
-    messages
-      .map(
-        ({ content, role }) => `<|im_start|>${role}
-${content}<|im_end|>`,
-      )
-      .join("\n") + "\n<|im_start|>assistant"
-  );
-}
-
-export function translateParams(
-  toProvider: ModelFormat,
-  params: Record<string, unknown>,
-  modelSpec?: ModelSpec | null,
-): Record<string, unknown> {
-  let translatedParams: Record<string, unknown> = {};
-
-  for (const [k, v] of Object.entries(params || {})) {
-    const safeValue = v ?? undefined; // Don't propagate "null" along
-    const translatedKey = modelParamToModelParam[k as keyof ModelParams] as
-      | keyof ModelParams
-      | undefined
-      | null;
-
-    if (translatedKey === null) {
-      continue;
-    }
-
-    const hasDefaultParam =
-      translatedKey !== undefined &&
-      Object.prototype.hasOwnProperty.call(
-        defaultModelParamSettings[toProvider] ?? {},
-        translatedKey,
-      );
-
-    translatedParams[hasDefaultParam ? translatedKey : k] = safeValue;
-  }
-
-  // ideally we should short circuit and just have a master mapper but this avoids scope
-  // for now
-  const mapper = paramMappers[toProvider];
-  if (mapper) {
-    translatedParams = mapper(
-      translatedParams as unknown as OpenAIChatCompletionCreateParams,
-      modelSpec,
-    ) as unknown as Record<string, unknown>;
-  }
-
-  return translatedParams;
-}
 
 export const anthropicSupportedMediaTypes = [
   "image/jpeg",
